3rd party tools
refactored tests

#TISNEW-3933
