+++ projects/hee-shared/src/lib/mock.test.component.ts
@@ -1,0 +1,20 @@
+/**
+ * MOCK DUMMY COMPONENTS AND ROUTES
+ */
+import { Component } from '@angular/core';
+import { Routes } from '@angular/router';
+
+@Component({
+    selector: 'app-mock',
+    template: `Mock`
+})
+export class MockComponent { }
+
+@Component({
+    template: `<router-outlet></router-outlet>`
+})
+export class AppComponent { }
+
+export const routes: Routes = [
+    { path: '', component: MockComponent }
+];
