+++ src/app/app.module.ts
@@ -5,7 +5,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { ServiceWorkerModule } from '@angular/service-worker';
 import { environment } from '@environment';
-import { AnalyticsModule, HotjarModule } from 'hee-shared';
+import { AnalyticsModule, HotJarModule } from 'hee-shared';
 
 @NgModule({
   declarations: [
@@ -16,7 +16,7 @@ import { AnalyticsModule, HotjarModule } from 'hee-shared';
     AppRoutingModule,
     ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production }),
     AnalyticsModule.forRoot({ siteId: ['UA-40570867-6'], enabled: environment.production }),
-    HotjarModule.forRoot({ hotjarid: 1662399, hotjarsv: 6, enabled: environment.production })
+    HotJarModule.forRoot({ hotJarId: 1662399, hotJarSv: 6, enabled: environment.production })
   ],
   providers: [],
   bootstrap: [AppComponent]
